requirements.txt
click==7.1.2
pdfminer.six==20200726
PyPDF2==1.26.0